Allow ignores list to contain git revisions
The ignores list in configs/ may now be a hash of ticket IDs to a list
of git revisions to permit more precise ignoring. When a ticket contains
multiple revisions from different projects, this prevents each revision
to be ignored without the danger of missing new commits against the
ticket.