Add deprecation warning scan
There is a related discussion here:
https://community.theforeman.org/t/deprecation-warnings-how-to-handle-better/17233/9

We will often forget to remove deprecation warnings in the release we say we will. This commit adds the parser gem and a class inheriting from ::Parser::AST::Processor which will scan for deprecation warning methods and any versions in the arguments. It the will compare them with the passed in version and print any outdated deprecation locations.