Swtich to centos7-updates
centos-7-server-updates had a republish of the openscap rpm, causing an issue with a metadata hash mishmatch.  Based upon my research, this is not fixable for the external repo it happened in, due to the way koji stores its information.  The fix is to use a "new" external repo that points to that location