Move manual updates to after tagging
The release notes are generated based on the content of Redmine. By
moving the manual updates after closing the Redmine version you're
guaranteed that the list of issues is in sync. Otherwise a change could
still go in. The titles could still be changes, but there's no way to
avoid that.

A hint is added that the optimal time to write is to wait while
packaging is building.